refactor: rename hook names
