+++ src/features/verida-database/hooks/use-delete-verida-record.ts
@@ -1,21 +1,22 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query"
 
 import { VeridaDatabaseQueryKeys } from "@/features/verida-database/queries"
-import { destroyVeridaDatabase } from "@/features/verida-database/utils"
+import { deleteVeridaDataRecord } from "@/features/verida-database/utils"
 import { useVerida } from "@/features/verida/hooks/use-verida"
 
-type DestroyVeridaDatabaseArgs = {
+type DeleteVeridaRecordArgs = {
   databaseName: string
+  recordId: string
 }
 
-export function useVeridaDestroyDatabase() {
-  const { getAccountSessionToken } = useVerida()
+export function useDeleteVeridaRecord() {
+  const { did, getAccountSessionToken } = useVerida()
   const queryClient = useQueryClient()
 
   const { mutate, mutateAsync, ...mutation } = useMutation({
-    mutationFn: async ({ databaseName }: DestroyVeridaDatabaseArgs) => {
+    mutationFn: async ({ databaseName, recordId }: DeleteVeridaRecordArgs) => {
       const sessionToken = await getAccountSessionToken()
-      return destroyVeridaDatabase({ databaseName, sessionToken })
+      return deleteVeridaDataRecord({ databaseName, recordId, sessionToken })
     },
     onSuccess: (_data, variables) => {
       queryClient.invalidateQueries({
@@ -24,20 +25,22 @@ export function useVeridaDestroyDatabase() {
         }),
       })
       queryClient.removeQueries({
-        queryKey: VeridaDatabaseQueryKeys.invalidateDataRecord({
+        queryKey: VeridaDatabaseQueryKeys.dataRecord({
+          did,
           databaseName: variables.databaseName,
+          recordId: variables.recordId,
         }),
       })
     },
     meta: {
       logCategory: "verida-database",
-      errorMessage: "Error destroying Verida database",
+      errorMessage: "Error deleting Verida data record",
     },
   })
 
   return {
-    destroyDatabase: mutate,
-    destroyDatabaseAsync: mutateAsync,
+    deleteRecord: mutate,
+    deleteRecordAsync: mutateAsync,
     ...mutation,
   }
 }
